+++ sdk/src/main/java/com/microsoft/did/sdk/backup/BackupParser.kt
@@ -1,20 +1,18 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.backup.container.jwe
+package com.microsoft.did.sdk.backup
 
 import com.microsoft.did.sdk.crypto.protocols.jose.jwe.JweToken
-import com.microsoft.did.sdk.datasource.backup.content.ProtectedBackupData
-import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
+import com.microsoft.did.sdk.backup.container.jwe.JwePasswordProtectedBackupData
+import com.microsoft.did.sdk.backup.content.ProtectedBackupData
+import com.microsoft.did.sdk.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
 import com.microsoft.did.sdk.util.controlflow.UnknownBackupFormatException
 import com.microsoft.did.sdk.util.controlflow.UnknownProtectionMethodException
-import kotlinx.serialization.json.Json
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class JweProtectedBackupFactory @Inject constructor(
-    private val jsonSerializer: Json
-) {
+class BackupParser @Inject constructor() {
     fun parseBackup(jweString: String): ProtectedBackupData {
         val token = JweToken.deserialize(jweString)
         val cty = token.contentType
