+++ sdk/src/main/java/com/microsoft/did/sdk/backup/container/ProtectionMethod.kt
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.backup.container
+package com.microsoft.did.sdk.backup.container
 
-import com.microsoft.did.sdk.datasource.backup.content.ProtectedBackupData
-import com.microsoft.did.sdk.datasource.backup.content.UnprotectedBackupData
+import com.microsoft.did.sdk.backup.content.ProtectedBackupData
+import com.microsoft.did.sdk.backup.content.UnprotectedBackupData
 import kotlinx.serialization.json.Json
 
 abstract class ProtectionMethod {
