+++ sdk/src/main/java/com/microsoft/did/sdk/backup/container/jwe/JwePasswordProtectedBackupData.kt
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.backup.container.jwe
+package com.microsoft.did.sdk.backup.container.jwe
 
 import com.microsoft.did.sdk.crypto.protocols.jose.jwe.JweToken
-import com.microsoft.did.sdk.datasource.backup.content.ProtectedBackupData
+import com.microsoft.did.sdk.backup.content.ProtectedBackupData
 
 class JwePasswordProtectedBackupData internal constructor(
     val jweToken: JweToken,
