+++ sdk/src/main/java/com/microsoft/did/sdk/backup/content/microsoft2020/Microsoft2020Backup.kt
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.backup.content.microsoft2020
+package com.microsoft.did.sdk.backup.content.microsoft2020
 
 import com.microsoft.did.sdk.credential.models.VerifiableCredential
-import com.microsoft.did.sdk.datasource.backup.UnprotectedBackup
+import com.microsoft.did.sdk.backup.UnprotectedBackup
 
 class Microsoft2020Backup(
     val walletMetadata: WalletMetadata,
