+++ sdk/src/test/java/com/microsoft/did/sdk/backup/ProtectedBackupDataFactoryTest.kt
@@ -1,12 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.file
+package com.microsoft.did.sdk.backup
 
-import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.VcMetadata
-import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.WalletMetadata
+import com.microsoft.did.sdk.backup.content.microsoft2020.VcMetadata
+import com.microsoft.did.sdk.backup.content.microsoft2020.WalletMetadata
 import android.util.VerifiableCredentialUtil
-import com.microsoft.did.sdk.datasource.backup.container.jwe.JweProtectedBackupFactory
-import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
+import com.microsoft.did.sdk.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
 import com.microsoft.did.sdk.util.defaultTestSerializer
 import org.junit.Test
 import java.io.ByteArrayInputStream
@@ -28,7 +27,7 @@ class ProtectedBackupDataFactoryTest {
         listOf(VerifiableCredentialUtil.rawIdentifier)
     )
 
-    val jweProtectedBackupFactory = JweProtectedBackupFactory(defaultTestSerializer)
+    val jweProtectedBackupFactory = BackupParser(defaultTestSerializer)
 
     @Test
     fun createPasswordBackupTest() {
