+++ sdk/src/test/java/com/microsoft/did/sdk/backup/content/microsoft2020/Microsoft2020UnprotectedBackupDataTest.kt
@@ -1,11 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.datasource.file
+package com.microsoft.did.sdk.backup.content.microsoft2020
 
 import android.util.VerifiableCredentialUtil
-import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.Microsoft2020UnprotectedBackupData
-import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.VcMetadata
-import com.microsoft.did.sdk.datasource.backup.content.microsoft2020.WalletMetadata
 import com.microsoft.did.sdk.di.defaultTestSerializer
 import org.junit.Test
 import kotlin.test.assertEquals
